relocates components
