+++ app/react/V2/Components/Forms/ConfirmNavigationModal.tsx
