+++ app/react/V2/Components/Forms/index.ts
@@ -1,5 +1,7 @@
 import { InputField } from './InputField';
 import { Select } from './Select';
 import { MultiSelect } from '../Forms/MultiSelect';
+import { FormInput } from '../Forms/FormInput';
+import { ConfirmNavigationModal } from '../Forms/ConfirmNavigationModal';
 
-export { InputField, Select, MultiSelect };
+export { InputField, FormInput, Select, MultiSelect, ConfirmNavigationModal };
