--- app/react/V2/Components/Translations/index.ts
@@ -1,4 +1,0 @@
-import { ConfirmNavigationModal } from './ConfirmNavigationModal';
-import { TranslationsTables } from './TranslationsTables';
-
-export { TranslationsTables, ConfirmNavigationModal };
