+++ app/react/V2/Routes/Settings/Translations/components/LanguagePill.tsx
@@ -1,5 +1,5 @@
+import { Pill } from 'app/V2/Components/UI';
 import React from 'react';
-import { Pill } from '../UI';
 
 const LanguagePill = ({ cell }: any) => {
   let color: 'gray' | 'primary' | 'yellow' = 'gray';
